--Workaround/Debugging Statement Updates--
code\Batch_DVM_execution.sql was updated to save the results of the SPT_QC_EVT_DIST_ISSUE_V as a materialized view of the same name so the slow query is not executed each time before evaluating the data validation criteria.

code\DVM_PKG.sql was updated to remove some verbose debugging statements.  Further development should be done to log statements in the database or a simple text file.